Fix flaky test TaskTest.inconsistentExecutionMode (#10031)
Summary:
The test is flaky because in parallel mode, the task might still be executing when the first scenario's scope is exited. Add waiting for task to finish at the end of the first scenario prevents the race condition to happen. Stress tests of 200 iterations have been run to make sure the fix is effective.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/10031

Reviewed By: xiaoxmeng

Differential Revision: D58115136

Pulled By: tanjialiang

fbshipit-source-id: 8ce4b30e73e785decbd53da26e2033ab66e97003