MergeJoin may hang if right side throws an exception (#9999)
Summary:

In MergeJoin if the left side is waiting for data (it's called next on the MergeJoinSource and there 
wasn't any available), it will create a promise consumerPromise_ to be fulfilled once the right side 
adds more data (it calls enqueue on the MergeJoinSource).

If the right side of the join fails before it can call enqueue, the Drivers will be closed and 
consumerPromise_ will be left unfulfilled, leaving the left side hanging.

To fix this, I fulfilled the promise in 2 places, one is in MergeJoinSource::close, the other is in 
enqueue when state.atEnd is already true.  It is safe to do so in either case because state.atEnd has 
already been set to true, so MergeJoinSource::next will return nullptr which is desired.

Either is sufficient to address the issue.  When the Driver is closed, first MergeJoinSource::close is 
called, then CallBackSink::close is called which will call MergeJoinSource::enqueue 
(MergeJoinSource::close will have already set state.atEnd to true).  I did it in both places to ensure
we don't miss any other edge cases.

Discovered running JoinFuzzer with OOM injection enabled.

Reviewed By: xiaoxmeng

Differential Revision: D58027361