Filters on LazyVectors in Joins can lead to incorrect results (#10045)
Summary:

A bug was recently discovered where joins with filters on the probe side can have incorrect results.

The root cause is that evaluating the filter will load LazyVectors, but will only load the rows needed to
evaluate the filter.  The subset of rows that are output may be a superset of this.

For example, a LEFT JOIN will always output all rows from the probe side.  If there's a filter on a 
column that is output with an "AND" that results in short circuiting only a subset of the rows will be 
loaded, and the column will be output with incorrect results for the unloaded rows.  (See the unit 
tests added in this change for clearer examples of when this can happen).

The fix is to pessimistically load LazyVectors from the probe side that are passed into the filter for
evaluation (similar to what we do when a single batch from the probe side crosses the boundaries of
output batches.

This bug appears to impact all join types except right semi filter and right semi project where the 
probe side is only used for filtering.

Reviewed By: Yuhta

Differential Revision: D58143920