Clean up some spurious shared_ptr creations/deletions in DwrfReader (#10095)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/10095

I was looking at a query reading from a table with many columns, and it's spending a huge amount
of CPU time just creating and deleting shared_ptrs when setting up the SplitReader and the
RowReader.

There's still more that can be done but this change cleans up some easy ones.

Differential Revision: D58263307