Guard metrics upload and report generation for fork PRs (#10080)
Summary:
I noticed in https://github.com/facebookincubator/velox/issues/9903 that the upload errors due to missing credentials. This PR adds a guard that allows the job to finish :heavy_check_mark: even from a fork. It will still run the upload (for testing) when the PR is from within the main repo.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/10080

Reviewed By: bikramSingh91

Differential Revision: D58244591

Pulled By: kevinwilfong

fbshipit-source-id: e75bf84292ad046dfc85be9f1a3b813af474cf11