Honor Presto's compression level configs and use the same defaults in HiveDataSink
Summary:
Presto allows configuring the compression level used with ZSTD and ZLIB compressors when writing
files to Hive.  This change updates HiveDataSink to apply the configured value to the compression
level it  uses.

It also updates the default value(s) used in HiveDataSink to match Presto.  For ZLIB this is 4 and for
ZSTD this is 3.  This was a little awkward given that they're controlled by a shared single config in
Presto.

Differential Revision: D58313487