Release pages to OS on contiguous allocations if evicted from cache
This applies to the MMAP allocator.
The Async data cache uses non-contiguous memory allocations to cached
data. If the allocator is running out of memory when attempting to
allocate more memory, it starts evicting pages from the cache.
These pages are deallocated but not released to the operating system in
anticipation that they will be reused by the allocation request.
However, if the allocation request is a large contiguous memory request and
it exceeds the allocator capacity, the cache starts evicting pages using the
freeNonContiguous function. This results in the size classes setting the deallocated
bit but keeps the pages mapped  but indicated as free (see  MmapAllocator::SizeClass::free()).

As a result the memory consumption by the allocator is reduced but from OS perspective
no memory was freed to be able to satisfy the contiguous mmap call which continues to fail
with not enough memory.

As a result of the continued failure the cache keeps deallocating memory without
resolving the underlying issue and eventually fails the query.

Thus, this PR attempts to free and release memory after the cache evicted
the requested amount of memory.
The MallocAllocator is not affected by this behavior.