Fix the flaky async cache tests
The flakiness is due to that currently we don't handle the concurrent cache
loads with cache shutdown. We rely on the query system to handle these concurrency
properly like stop accessing query cache access, and then shutdown the cache.
This is not the case for async cache tests which might schedule coalesced cache
loads to a background executor, and might start the next round of tests without
waiting for the previous background activities to finishes. Each test round like
in ssdWriteOptions will reset the cache which shutdown the cache. This cause the
background coalesced cache load see freed cache entries. This change fixes the flakines
by adding a pending batch load counter and wait for it to complete before starts
the next round of test.

Note we have no plan to handle the async cache shutdown race with concurrent query
cache access. It might need a lock which is pretty expensive as in-memory cache access
is high frequent operation and we would rather relying on the query system to prevent
this.