Lazy materialization of RowType::parameters (#10144)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/10144

As part of optimization for reading tables with large number of columns
(> 7000), the materialization of `RowType::parameters` of table/file schema
costs a non-trivial amount of CPU and the parameters are never used.  Making it
lazy saves large amount of CPU for these tables.

As a result of laziness, `RowType` copy constructor is deleted.  This helps us
spotting some places in code base where we copy the object unnessarily and this
change cleans them up.

Reviewed By: oerling

Differential Revision: D58417885