Keep DuckDB re2 headers private (#10161)
Summary:
DuckDB source when bundled includes a copy of re2 as a third-party dependency. However, it defines
the re2 headers as public which seem to conflict with the system re2 headers.

The base build needs the manual install_duckdb since it is being built off master.

Developers must delete the `_deps/duckdb_*` folders in their build folder for the patches to apply cleanly.

Resolves https://github.com/facebookincubator/velox/issues/10154

Pull Request resolved: https://github.com/facebookincubator/velox/pull/10161

Reviewed By: zacw7

Differential Revision: D58473939

Pulled By: kagamiori

fbshipit-source-id: d924ca749eb7103d61e4b16dd77d46ef642c7dd0