Prevent cache write from exceeding IOV_MAX (#10142)
Summary:
When calling `pwritev` with `iovecs` sized larger than `IOV_MAX`, the
call will fail, resulting in an SSD write error. To prevent this, it is
necessary to write out the accumulated `iovecs` before they exceed the
IOV_MAX limit.
This change was previously introduced in commit 8d09dc9 but contained a
bug: the number of bytes to write is calculated by `getSpace()` and
stored in `available`. The for loop relies on the condition
`writeLength + entrySize > available` to determine when to stop writing.
However, each time it writes out the accumulated iovecs, `writeLength` is
reset to zero, making the condition unreliable and sometimes causing it
to write more than intended.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/10142

Reviewed By: xiaoxmeng

Differential Revision: D58430846

Pulled By: zacw7

fbshipit-source-id: 0f9720f7a082e2a98d685334415aa9caca5ec185