Verify checksum for full SSD entry only
When loading data from 'ssdPins' into 'pins', mismatches occur if the
requested 'pins' range is shorter than the SSD cache entries. This is
because the checksum is calculated on partial data but compared with a
checksum of the full entry. To resolve this, checksum verification
should only be applied when reading the entire SSD entry.