Refactor timestamp with timezone unit tests
Summary:
Refactoring some of the unit tests using the timestamp with time zone data
type to make them more concise and consistent with other tests by using
"evaluateOnce()", in addition to allowing it to tests null input and output.
Also adding some more convenience functions to the test helper class
TimestampWithTimezone.

Looking into this since it was pretty hard to understand and compare output of
date_add() unit tests, since they were done based on the packed values.

There is still *a lot* of clean up to be done in this unit test suite. Fixing a
few callsites for now, but leaving additional callsites to a future refactor.

Differential Revision: D58471534