Fix autodeps for xsimd
Summary:
Fix the autodeps for xsimd.  Currently auto deps tries to pull in pythran if a file depends on xsimd.
This is wrong and leads to broken builds if we just blindly accept it.

Fix autdeps to pull in xsimd from third-party if code depends on a header from xsimd.

Currently Velox is the only project that uses xsimd directly (some other projects use it as well, but
pull in Velox's header files to do so).

This also allows us to remove a buck directive from our open source Velox code.

Differential Revision: D58592995