Implement bucket conversion in HiveDataSource (#10228)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/10228

When the bucket count of a table changes over time, there can be
legitimate cases that multiple buckets exist in the same file.  In such cases
the query planner should set bucket conversion for these splits and in
`HiveDataSource` we add extra filter to get only the rows corresponding to the
bucket number requested.

Also clean up a little bit of the row number generation code and add some sanity checks.

Differential Revision: D58670760