Add cmake helper to find libdouble-conversion
The package of libdouble-conversion-dev in Ubuntu 20.04 does not provide
a cmake config file. Therefore, cmakes fails to initialize stating that
it cannot find any of the names:
  double-conversionConfig.cmake,
  double-conversion-config.cmake.

This change provides a cmake helper to let Velox find the library.
It was originally copied from:
  https://github.com/qt/qtbase/blob/v6.3.1/cmake/Finddouble-conversion.cmake
that was later refactored into:
  https://github.com/qt/qtbase/blob/dev/cmake/FindWrapSystemDoubleConversion.cmake

Fixes #10230