Fix flaky test TableScanTest.statsBasedSkippingNulls
Summary:
The test is flaky in optimized build on better hardware because the
minimum unit of our timer is 1 microsecond, so if one read is faster than it,
the scan time could be recorded as 0.  Remove the flaky check and beef up the
test with some stats it should check against.

Differential Revision: D58691228