Fix Velox counts incomplete UTF-8 chars differently than Presto Java (#10204)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/10204

In String UDFs, when Presto is counting the number of code points in a string, e.g. to compute its
length or to compute a substring it simply ignores UTF-8 continuation characters.

Velox assumes that if it encounters a UTF-8 prefix byte the rest of the bytes in the UTF-8 code
point are well formed (i.e. they're all continuation bytes).  This causes Velox and Presto to count
code points differently.

While both approaches produce (IMO) odd results at times, Velox's seems to be clearly worse (e.g.
a UTF-8 code point with a valid prefix indicating 2 continuation bytes but those continuation bytes
are missing will effectively count as -2 when computing the length) and we usually default to
Presto Java as the source of truth, so I've updated Velox.

Reviewed By: kgpai

Differential Revision: D58620186