Fix parallel spills lead to crashes in approx_percentile (#10268)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/10268

A recent optimization introduced the ability to spill multiple partitions of an aggregation in parallel.

This has caused queries that use the approx_percentile aggregation to crash, frequently the crash
points to the HashStringAllocator being used to allocate memory in extractAccumulators and finding
that the header it received from the freeList was not marked free.

The root cause is a concurrency issue.  HashStringAllocator is not thread safe.  The KllSketch class
allocates memory while compacting the KllSketch as well as when the KllSketchAccumulator merges
the largeCountValues_ into the KllSketch, both of which happen in KllSketchAccumulator's finalize
method which is invoked by extractAccumulators as part of spilling.  When two KllSketch's try to
get a free header concurrently, they may both find the same one, and then only one wins marking
the header as not free (it's actually possible that both "win" which leads to memory corruption and
crashes elsewhere).

The fix is to copy the KllSketch into a new one which uses an unshared stl::allocator (thread-safe)
instead of the shared HashStringAllocator, and then do the merging and compaction.  There is some
additional cost of copying some std::vectors inside of KllSketch (levels and items), but it's similar to
the cost we pay anyway for growing the vectors beyond their capacity which involves a copy as well.

The extractValues method in ApproxPercentileAggregate currently uses the same logic, but is not
invoked in a thread-unsafe manner, I also found it's unnecessarily compacting the KllSketch.  I
modified it to call flush on the KllSketchAccumulator which only merges the largeCountValues_ and
calls finish (rather than compact) on the KllSketch, all within the original KllSketch (no additional
copy).

To do this I had to move the inner class View out of KllSketch into a top level class, since KllSketch is
templated on the Allocator, but the View doesn't actually depend on it (this makes copying from a
View over a KllSketch with one Allocator into a KllSketch with another Allocator easy).

Reviewed By: Yuhta

Differential Revision: D58823050