Fix race condition in MemoryArbitrationFuzzer
Summary:
MemoryArbitrationFuzzer reuses the same rng in multiple threads, this leads to TSAN reporting a
data race.  We can use a distinct rng in each thread with a seed generated by the original rng to
avoid this.

(The tests aren't entirely reproducible since they depend on the timing of the threads, so using
seeds from the original rng is not really necessary)

Differential Revision: D59031526