Fix race conditions in AsyncDataCache AccessStats (#10312)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/10312

AsyncDataCache uses an AccessStats object to track when an entry was last accessed and how
many times it's been accessed.  With multiple threads accessing entries and threads that may have
never used an entry checking to see if it can be evicted from the cache to make space, these need
to be guarded to ensure consistency in the face of concurrent updates/reads.

This was exposed running CacheTest.readAhead with TSAN enabled.

Differential Revision: D59021411