Ensure the shared HashStringAllocator isn't mutated by Aggregations during spilling (#10309)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/10309

In https://github.com/facebookincubator/velox/pull/10268 we discovered and fixed an issue where
approx_percentile was running into concurrency issues in HashStringAllocator during spilling
because parallel spiling would lead to multiple concurrent allocations from the same
HashStringAllocator.

This was fixed for approx_percentile, but there may still be issues in other aggregates (existing or
future).  To catch these clearly and early I've add a mechanism to freeze HashStringAllocator.

When HashStringAllocator is frozen any attempt to access any of its fields in a mutable way will
result in an exception.  This is achieved by wrapping its fields in an inner class State which exposes
them via getter methods that can either return a const or non-const reference.  Before returning a
non-const reference it first checks if the class is frozen.

I've updated the places in GroupingSet where we invoke spilling to freeze the HashStringAllocator
while this process is happening.  This way if any Aggregation may trigger a concurrency issue it is
detected and an exception is thrown.

This relies on future developers following the pattern of adding fields to the State class with
accessors, I've added comments to do my best to ensure they follow this pattern.

Reviewed By: xiaoxmeng

Differential Revision: D59012548