Reduce array_filter and map_filter memory usage
Summary:
We observed some OOM happening because the downstream operator of
`map_filter` is accumulating the input vectors.  These vectors are dictionary
with very few indices but a large alphabet, i.e. result of filtering.  To
prevent OOM from such cases, we add in a threshold that when the filtering ratio
is below it, we copy the result instead of holding on to a large base vector to
reduce memory usage.

Differential Revision: D59125367