Fix the initial cache stats reporting (#10408)
Summary:
There is a bug in delta cache stats calculation which can leads to the missing initial report.
As we only set the ssd delta if both the old and new cache stats has ssd stats but for the
initial report, the old one doesn't have the ssd stats which caused the initial cache stats like
the checkpoint reads never get reported. This PR fixes this with unit test.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/10408

Reviewed By: tanjialiang, zacw7

Differential Revision: D59416042

Pulled By: xiaoxmeng