Fix race condition in CacheFuzzer
Summary:
The threads CacheFuzzer spins up use the loop variable "i" captured by reference as part of
generating the seed.  This loop variable changes on every iteration so we get concurrent reads and
writes to this variable which TSAN captures (and also probably violates the intention of the author).

Capturing a copy of the variable fixes it.

Differential Revision: D59476752