Rewrite some common reduce patterns with array_sum and transform (#10406)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/10406

Lambda expression in Velox is interpreted instead of compiled like in
Presto Java, this makes us losing in performance in places like `reduce`.  By
rewriting some common patterns of `reduce`, we improve the performance
significantly (more than 45x on a query).

Also fix a bug in `TryExpr` where we are not checking enough space for result
nulls.

Reviewed By: oerling

Differential Revision: D59406443