Fix improper template instantiation in ColumnWriterTest
The template instantiation when using complex types
results in attempting to create a vector that is only
implemented for use with simple types.

This can be seen when attempting to compile with Clang 17.
For example the following is attempted to be instantiated:
facebook::velox::SimpleVector<facebook::velox::Array<int>>::comparePrimitiveAsc