Fix the timing order of DeltaCpuWallTimer
A counter-intuitive phenomenon is often observed: in the `addInput`
and `getOutput` timing metrics of some operators, cpu-time is slightly
larger than wall-time.

It is intuitive that wall-time should be slightly larger than cpu-time.

The `addInput` and `getOutput` timing metrics of the operators are
automatically counted using `DeltaCpuWallTimer`, including cpu-time and
wall-time. In order to correctly count the timing, the correct order
should be:

```
1. wall-time starts counting;
2. cpu-time starts counting;
3. cpu-time ends counting: delta value is the cpu-time of one OP.
4. wall-time ends counting: delta value is the wall-time of one OP;
```

That is, the timing of wall-time should start before cpu-time and end
after cpu-time.

However, the current implementation of `DeltaCpuWallTimer` does not
follow this order. First, it will initialize `cpuTimeStart_`, because
the declaration order of the `cpuTimeStart_` member is before
`wallTimeStart_`, so it will initialize `cpuTimeStart_` first.
Secondly, in the destructor of `DeltaCpuWallTimer`, the current order is
to end the wall-time timing first, and then end the cpu-time timing.

This PR corrects the above counting order in `DeltaCpuWallTimer`.