Use HashStringAllocator for raw_vector rows in HashLookup
The rows in the HashLookup are stored as raw_vector.
The raw_vector uses a malloc type allocation to acquire aligned
memory. It was found this contributes to large unaccounted
memory when  mainly using the HashProbe and other operators.

This change allows the usage of a custom allocator for the
raw_vector. By default it keeps the current way of allocating
and deallocating.

For raw_vectors representing rows in the HashLookup the
allocator used is the HashStringAllocator which allows for
the accounting of the memory.

Co-authored-by: Jimmy Lu <gongchuo.lu@gmail.com>