Add support for PrestoQueryRunner in AggregationRunnerTest.cpp (#10498)
Summary:
AggregationRunnerTest serves as a handy tool for replaying, debugging, and validating failures detected by the aggregation fuzzer. However, it currently lacks support for using Presto as the source of truth (SoT), despite the fact that we use Presto as SoT when running the Aggregation Fuzzer. This pull request addresses and resolves this gap.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/10498

Reviewed By: kagamiori

Differential Revision: D59884675

Pulled By: kgpai