Timezone conversions based on seconds
Summary:
Timezone conversions were done based on a timepoint<milliseconds>,
which is unecessary, adds possibility for overflow, and requires us to floor it
after the conversion. Simplifying the code and adding some more tests.

Now, only exporting a time_point<millisecond> may overflow, but not exporting
to time_point<second> which is what is needed for timezone conversions.

Reviewed By: mbasmanova

Differential Revision: D60189530